+++ test/template-fragment.tsx
@@ -1,0 +1,8 @@
+import * as React from "preact";
+
+export default function render() {
+  return <>
+      <div>Hello</div>
+      <div>world!</div>
+    </>;
+}