+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+* Drop support for wrapper tags (templates should use fragments instead)
+
 ## 0.3.2 (July 21, 2020)
 
 * Downgrade fs-extra to 8.1.0 for better Node compatibility